[SYCL][COMPAT] shared memory test split from main tests
